requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
requests>=2.31.0
lxml>=5.0.0
html5lib>=1.1
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.126.0
